Refactor js code for mobile- and desktop-specific behavior
Previously, the viewport width is only checked upon page load. If the
user resizes the viewport, say from desktop to mobile layout, then the
dropdown menus still have hover interactivity, unless the page is
reloaded in the mobile layout.

This is fixed in this commit by moving the viewport logic inside the
callout functions, so that every time the user hovers over the dropdown
menu, the viewport width value is refreshed, and if the user sees the
desktop layout, then the hover functionality is activated.

Same for the mobile layout.

The drawback is slightly more repetitive code, but this can be improved
in the future.